backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
python-docx==0.8.11
PyPDF2==3.0.1
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.1